Move prepareStepLibForActivation into ActivateStep, return ResolvedStep
ActivateStep now takes a stepid.CanonicalID and owns step resolution: it
calls prepareStepLibForActivation (moved here from the activator package)
and activates the resolved stepInfo directly, dropping the redundant
queryStepMetadata re-resolution. It returns a ResolvedStep carrying the
optional executable path, the resolved StepInfo, and whether the local
steplib cache was updated.

The obsolete activator-package resolution test is removed here; its
replacement lands next to the moved function in a follow-up commit.
