+++ activator/steplib_ref.go
@@ -1,12 +1,9 @@
 package activator
 
 import (
-	"fmt"
 	"path/filepath"
 
-	"github.com/bitrise-io/go-utils/pointers"
 	"github.com/bitrise-io/stepman/activator/steplib"
-	"github.com/bitrise-io/stepman/models"
 	"github.com/bitrise-io/stepman/stepid"
 	"github.com/bitrise-io/stepman/stepman"
 )
@@ -26,16 +23,11 @@ func ActivateSteplibRefStep(
 		DidStepLibUpdate: false,
 	}
 
-	stepInfo, didUpdate, err := prepareStepLibForActivation(log, id, didStepLibUpdateInWorkflow, isOfflineMode)
-	activationResult.DidStepLibUpdate = didUpdate
-	activationResult.StepInfo = stepInfo
-	if err != nil {
-		return activationResult, err
-	}
-
-	execPath, err := steplib.ActivateStep(id.SteplibSource, id.IDorURI, stepInfo.Version, activatedStepDir, stepYMLPath, log, isOfflineMode)
-	activationResult.ExecutablePath = execPath
-	if execPath != "" {
+	resolvedStep, err := steplib.ActivateStep(id, activatedStepDir, stepYMLPath, log, didStepLibUpdateInWorkflow, isOfflineMode)
+	activationResult.StepInfo = resolvedStep.StepInfo
+	activationResult.DidStepLibUpdate = resolvedStep.DidStepLibUpdate
+	activationResult.ExecutablePath = resolvedStep.ExecPath
+	if resolvedStep.ExecPath != "" {
 		activationResult.ActivationType = ActivationTypeSteplibExecutable
 	} else {
 		activationResult.ActivationType = ActivationTypeSteplibSource
@@ -45,83 +37,4 @@ func ActivateSteplibRefStep(
 	}
 
 	return activationResult, nil
-}
-
-func prepareStepLibForActivation(
-	log stepman.Logger,
-	id stepid.CanonicalID,
-	didStepLibUpdateInWorkflow bool,
-	isOfflineMode bool,
-) (stepInfo models.StepInfoModel, didUpdate bool, err error) {
-	err = stepman.SetupLibrary(id.SteplibSource, log)
-	if err != nil {
-		return models.StepInfoModel{}, false, fmt.Errorf("setup %s: %s", id.SteplibSource, err)
-	}
-
-	versionConstraint, err := models.ParseRequiredVersion(id.Version)
-	if err != nil {
-		return models.StepInfoModel{}, false, err
-	}
-	if versionConstraint.VersionLockType == models.InvalidVersionConstraint {
-		return models.StepInfoModel{}, false, fmt.Errorf("version constraint is invalid: %s %s", id.IDorURI, id.Version)
-	}
-
-	if shouldUpdateStepLibForStep(versionConstraint, isOfflineMode, didStepLibUpdateInWorkflow) {
-		log.Infof("Step uses latest version, updating StepLib...")
-		_, err = stepman.UpdateLibrary(id.SteplibSource, log)
-		if err != nil {
-			log.Warnf("Step version constraint is latest or version locked, but failed to update StepLib, err: %s", err)
-		} else {
-			didUpdate = true
-		}
-	}
-
-	stepInfo, err = stepman.QueryStepInfoFromLibrary(id.SteplibSource, id.IDorURI, id.Version, log)
-	if err != nil {
-		if !canUpdateStepLib(isOfflineMode, didStepLibUpdateInWorkflow) {
-			return stepInfo, didUpdate, err
-		}
-
-		log.Infof("Step not found in local StepLib cache, trying to update StepLib...")
-		_, err = stepman.UpdateLibrary(id.SteplibSource, log)
-		if err != nil {
-			return stepInfo, didUpdate, err
-		} else {
-			didUpdate = true
-		}
-
-		stepInfo, err = stepman.QueryStepInfoFromLibrary(id.SteplibSource, id.IDorURI, id.Version, log)
-		if err != nil {
-			return stepInfo, didUpdate, err
-		}
-	}
-
-	if stepInfo.Step.Title == nil || *stepInfo.Step.Title == "" {
-		stepInfo.Step.Title = pointers.NewStringPtr(stepInfo.ID)
-	}
-	stepInfo.OriginalVersion = id.Version
-
-	return stepInfo, didUpdate, nil
-}
-
-func shouldUpdateStepLibForStep(constraint models.VersionConstraint, isOfflineMode bool, didStepLibUpdateInWorkflow bool) bool {
-	if !canUpdateStepLib(isOfflineMode, didStepLibUpdateInWorkflow) {
-		return false
-	}
-
-	return (constraint.VersionLockType == models.Latest) ||
-		(constraint.VersionLockType == models.MinorLocked) ||
-		(constraint.VersionLockType == models.MajorLocked)
-}
-
-func canUpdateStepLib(isOfflineMode bool, didStepLibUpdateInWorkflow bool) bool {
-	if isOfflineMode {
-		return false
-	}
-
-	if didStepLibUpdateInWorkflow {
-		return false
-	}
-
-	return true
 }
