+++ activator/steplib_ref_test.go
@@ -10,108 +10,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func Test_activateStepLibStep(t *testing.T) {
-	tests := []struct {
-		name        string
-		stepIDData  stepid.CanonicalID
-		wantVersion string
-		wantErr     bool
-	}{
-		{
-			name: "Major version lock",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "1",
-			},
-			wantVersion: "1.10.1",
-			wantErr:     false,
-		},
-		{
-			name: "Major version lock (long form)",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "1.x.x",
-			},
-			wantVersion: "1.10.1",
-			wantErr:     false,
-		},
-		{
-			name: "Minor version lock",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "2.3",
-			},
-			wantVersion: "2.3.7",
-			wantErr:     false,
-		},
-		{
-			name: "Minor version lock (long form)",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "2.3.x",
-			},
-			wantVersion: "2.3.7",
-			wantErr:     false,
-		},
-		{
-			name: "Patch version lock",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "2.3.2",
-			},
-			wantVersion: "2.3.2",
-			wantErr:     false,
-		},
-		{
-			name: "Invalid version lock",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "1.2.3.4",
-			},
-			wantVersion: "",
-			wantErr:     true,
-		},
-		{
-			name: "Latest version (not supported at the moment)",
-			stepIDData: stepid.CanonicalID{
-				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
-				IDorURI:       "xcode-archive",
-				Version:       "x.x.x",
-			},
-			wantVersion: "",
-			wantErr:     true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tmpDir, err := pathutil.NormalizedOSTempDirPath("activateStepLibStep")
-			if err != nil {
-				t.Errorf("failed to create tmp dir: %s", err)
-			}
-			stepYMLCopyPth := filepath.Join(tmpDir, "step-yml", "step.yml")
-
-			if err := os.MkdirAll(filepath.Dir(stepYMLCopyPth), 0777); err != nil {
-				t.Errorf("failed to create dir for step.yml: %s", err)
-			}
-
-			got, _, err := prepareStepLibForActivation(TestLogger[*testing.T]{t}, tt.stepIDData, false, false)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("activateStepLibStep() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got.Version != tt.wantVersion {
-				t.Errorf("activateStepLibStep() got = %v, want %v", got.Version, tt.wantVersion)
-			}
-		})
-	}
-}
-
 type genericLogger interface {
 	Logf(format string, v ...any)
 }
